comments and typos

+++ src/01_VariableEtMemoire/src/example_cast.cpp
@@ -22,7 +22,7 @@ int main(){
     /// [static_overflow]
     int value = std::numeric_limits<short int>::max()+10;
     std::cout << "value: " << value << std::endl;
-    float sivalue = static_cast<short int>(value); 
+    short int sivalue = static_cast<short int>(value); 
     std::cout << "sivalue: " << sivalue << std::endl;
     /// [static_overflow]
     }
